registration model for kernel finder
