+++ src/kernels/raw/types.ts
@@ -6,8 +6,8 @@
 import { CancellationToken, Event } from 'vscode';
 import { IAsyncDisposable, IDisplayOptions, IDisposable, Resource } from '../../platform/common/types';
 import {
+    IContributedKernelFinder,
     IKernelConnectionSession,
-    INotebookProviderConnection,
     KernelConnectionMetadata,
     LocalKernelConnectionMetadata,
     LocalKernelSpecConnectionMetadata,
@@ -58,21 +58,13 @@ export interface IKernelProcess extends IDisposable {
 }
 
 export const ILocalKernelFinder = Symbol('ILocalKernelFinder');
-export interface ILocalKernelFinder {
+export interface ILocalKernelFinder extends IContributedKernelFinder {
     /**
      * Finds all kernel specs including Python.
      */
     listKernels(resource: Resource, cancelToken?: CancellationToken): Promise<LocalKernelConnectionMetadata[]>;
 }
 
-export const IRemoteKernelFinder = Symbol('IRemoteKernelFinder');
-export interface IRemoteKernelFinder {
-    listKernels(
-        resource: Resource,
-        connInfo: INotebookProviderConnection | undefined,
-        cancelToken?: CancellationToken
-    ): Promise<KernelConnectionMetadata[]>;
-}
 /**
  * The daemon responsible for the Python Kernel.
  */
